pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
authors = [
  {email = "thomas@gibfest.dk"},
  {name = "Thomas Steen Rasmussen"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "dealer==2.1.0",
    "Django==5.0.3",
    "django-allauth==0.61.1",
    "django-bootstrap5==23.4",
    "django-cors-headers==4.3.1",
    "django-guardian==2.4.0",
    "django-htmx==1.17.3",
    "django-imagekit==5.0.0",
    "django-ninja==1.1.0",
    "django-oauth-toolkit==2.3.0",
    "django-polymorphic==3.1.0",
    "django-taggit==5.0.1",
    "environs[django]==11.0.0",
    "fontawesomefree==6.5.1",
    "orjson==3.9.15",
    "Pillow==10.2.0",
    "psycopg2-binary==2.9.9",
    "python-magic==0.4.27",
]
description = "BornHack Media Archive Django project"
name = "bma"
version = "0.1"
readme = "README.md"
requires-python = ">=3.10"

[project.scripts]
"manage.py" = "bma.manage:main"

[project.optional-dependencies]
dev = [
    "pre-commit==3.3.2",
    "django-debug-toolbar==4.1.0",
]
test = [
    "coverage==7.2.2",
    "factory-boy==3.2.1",
]

[project.urls]
homepage = "https://github.com/bornhack/bma"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
extend-exclude = [
    ".git",
    "__pycache__",
    "*/migrations/*.py"
]
lint.select = ["ALL"]
lint.ignore = [
    "G004", # https://docs.astral.sh/ruff/rules/logging-f-string/
    "ANN101", # https://docs.astral.sh/ruff/rules/missing-type-self/
    "ANN102", # https://docs.astral.sh/ruff/rules/missing-type-cls/
    "EM101", # https://docs.astral.sh/ruff/rules/raw-string-in-exception/
    "EM102", # https://docs.astral.sh/ruff/rules/f-string-in-exception/
    "COM812", # missing-trailing-comma (https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules)
    "ISC001", # single-line-implicit-string-concatenation (https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules)
    "ARG001", # https://docs.astral.sh/ruff/rules/unused-function-argument/
    "ARG002", # https://docs.astral.sh/ruff/rules/unused-method-argument/
    "ARG004", # https://docs.astral.sh/ruff/rules/unused-static-method-argument/
]
line-length = 120

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests.py" = [
    "S101", # https://docs.astral.sh/ruff/rules/assert/
    "PLR2004", # https://docs.astral.sh/ruff/rules/magic-value-comparison/
]
"*/migrations/*" = [
    "D" # https://docs.astral.sh/ruff/rules/#pydocstyle-d
]
